misc/HtmlTestSuite/results/test-0382.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><title>print-color-adjust Parsing</title>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>

<link href="https://www.w3.org/TR/css-color-adjust-1/#perf" rel="help"/>
<link href="https://www.xanthir.com/contact/" rel="author" title="Tab Atkins-Bittner"/>

<script src="/css/support/parsing-testcommon.js"/>

<script>
test_valid_value('print-color-adjust', 'exact', 'exact');
test_valid_value('print-color-adjust', 'economy', 'economy');
test_valid_value('color-adjust', 'exact', 'exact');
test_valid_value('color-adjust', 'economy', 'economy');
</script>
</head><body/></html>
